Experiments that we have done for Sensor Yourself
In this video we have presented a demonstration of how the thermocouple
works. The two wires (made of different materials) are interlinked with
each other. On one end the temperature is the same as the room
temperature whereas the other end is heated by a gas cylinder whose
temperature can go up to 3600 Fahrenheit degrees. Due to the movement of
electrons a voltage (therefore a current) is generated, which we can
clearly see on the multimeter display.